Taxgig v1.0.0-beta.1-0-g3b4f48f+dev Chat.Token

Custom token implementation using EntropyString behaviour and Phoenix Token.

Functions

bits()

Default entropy bits for random strings

chars()

Module characters

charset()

Module EntropyString.CharSet

ets_token()

flip(flips)

large()

Random string using module charset with a 1 in a trillion chance of repeat for a billion potential strings.

Example

MyModule.large()
"NqJLbG8htr4t64TQmRDB"

medium()

Random string using module charset with a 1 in a billion chance of repeat for a million potential strings.

Example

MyModulue.medium()
"nndQjL7FLR9pDd"

random(bits \\ 128)

Random string of entropy bits using module charset

bits - entropy bits for string

non-negative integer

predefined atom

Defaults to module bits

Returns string of at least entropy bits using module characters; or

{:error, "Negative entropy"} if bits is negative.

{:error, reason} if EntropyString.CharSet.validate(charset) is not true.

Since the generated random strings carry an entropy that is a multiple of the bits per module characters, the returned entropy is the minimum that equals or exceeds the specified bits.

Example

A million potential strings (assuming :charset32 characters) with a 1 in a billion chance of a repeat

bits = EntropyString.bits(1.0e6, 1.0e9)

MyModule.random(bits)
"NbMbLrj9fBbQP6"

MyModule.random(:session)
"CeElDdo7HnNDuiWwlFPPq0"

random_string(len)

session()

Random string using module charset suitable for 128-bit OWASP Session ID

Example

MyModule.session()
"6pLfLgfL8MgTn7tQDN8tqPFR4b"

small()

Random string using module charset with a 1 in a million chance of repeat in 30 strings.

Example

MyModule.small()
"nGrqnt"

string()

Random string of module entropy bits and charset

Example

Define a module for 10 billion strings with a 1 in a decillion chance of a repeat

defmodule Rare, do: use EntropyString, total: 1.0e10, risk: 1.0e33

Rare.string()
"H2Mp8MPT7F3Pp2bmHm"

Define a module for strings with 122 bits of entropy

defmodule MyId, do: use EntropyString, bits: 122, charset: charset64

MyId.string()
"aj2_kMH64P2QDRBlOkz7Z"

string_of_length(length)

token()

Random string using module charset with 256 bits of entropy.

Example

MyModule.token()
"zHZ278Pv_GaOsmRYdBIR5uO8Tt0OWSESZbVuQye6grt"