Taxgig v1.0.0-beta.1-0-g3b4f48f+dev Core.Upload

Manage user uploads

Options:

:type: presets for activity type (defaults to Document) and size limits from app configuration

:description: upload alternative text

:base_url: override base url

:uploader: override uploader

:filters: override filters

:size_limit: override size limit

:activity_type: override activity type

The %Core.Upload{} struct: all documented fields are meant to be overwritten in filters:

:id - the upload id.

:name - the upload file name.

:path - the upload path: set at first to id/name but can be changed. Keep in mind that the path is once created permanent and changing it (especially in uploaders) is probably a bad idea!

:tempfile - path to the temporary file. Prefer in-place changes on the file rather than changing the path as the temporary file is also tracked by Plug.Upload{} and automatically deleted once the request is over.

Related behaviors:

Core.Uploaders.Uploader

Core.Upload.Filter

Types

option()

Specs

option() ::
  {:type, :avatar | :banner | :background}
  | {:description, String.t()}
  | {:activity_type, String.t()}
  | {:size_limit, nil | non_neg_integer()}
  | {:uploader, module()}
  | {:filters, [module()]}

source()

Specs

source() ::
  Plug.Upload.t()
  | (data_uri_string :: String.t())
  | {:from_local, name :: String.t(), id :: String.t(), path :: String.t()}

t()

Specs

t() :: %Core.Upload{
  content_type: String.t(),
  id: String.t(),
  name: String.t(),
  path: String.t(),
  size: integer(),
  tempfile: String.t()
}

Functions

char_unescaped?(char)

Specs

char_unescaped?(char()) :: boolean()

remove(url, opts \\ [])

Specs

remove(String.t(), list()) :: String.t() | {:error, String.t()}

store(upload, opts \\ [])

Specs

store(source(), [option()]) :: {:ok, map()} | {:error, any()}