Taxgig v1.0.0-beta.1-0-gede5592+dev Core.Uploaders.Uploader behaviour

Defines the contract to put and get an uploaded file to any backend.

Types

file_spec()

file_spec() :: {:file | :url, String.t()}

get_method()

get_method() ::
  {:static_dir, directory :: String.t()} | {:url, url :: String.t()}

Functions

put_file(uploader, upload)

put_file(module(), Core.Upload.t()) :: {:ok, file_spec()} | {:error, String.t()}

remove_file(uploader, path)

remove_file(module(), String.t()) ::
  :ok | {:ok, file_spec()} | {:error, String.t()}

Callbacks

get_file(file)

get_file(file :: String.t()) :: {:ok, get_method()}

Instructs how to get the file from the backend.

Used by Core.Plugs.UploadedMedia.

http_callback(arg1, arg2)
(optional)

http_callback(Plug.Conn.t(), Map.t()) ::
  {:ok, Plug.Conn.t()}
  | {:ok, Plug.Conn.t(), file_spec()}
  | {:error, Plug.Conn.t(), String.t()}

put_file(arg1)

put_file(Core.Upload.t()) ::
  :ok | {:ok, file_spec()} | {:error, String.t()} | :wait_callback

Put a file to the backend.

Returns:

:ok which assumes {:ok, upload.path}

{:ok, spec} where spec is: {:file, filename :: String.t} to handle reads with get_file/1 (recommended)

This allows to correctly proxy or redirect requests to the backend, while allowing to migrate backends without breaking any URL.

{url, url :: String.t} to bypass get_file/2 and use the url directly in the activity.

{:error, String.t} error information if the file failed to be saved to the backend.

:wait_callback will wait for an http post request at /api/community/upload_callback/:upload_path and call the uploader's http_callback/3 method.

remove_file(file_spec)

remove_file(file_spec()) :: :ok | {:ok, file_spec()} | {:error, String.t()}