Taxgig v1.0.0-beta.1-0-gede5592+dev Mailings.Mailer

A client API to the Mailgun .

Functions

conf()

send_email(email)

send_forgot_password_html(id, email, name)

send_forgot_password_html(String.t(), String.t(), String.t()) :: :ok

Send an email in format html if customer forgot password.

Examples

iex> Mailings.Mailer.send_forgot_password_html("123", "test@example.com", "John")
:ok

send_pro_html(customer)

send_pro_html(String.t()) :: :ok

Send an email in format html for customer PRO.

Examples

iex> Mailings.Mailer.send_pro_html("test@example.com")
:ok

send_pro_text(customer)

send_pro_text(String.t()) :: :ok

Send an email in format txt for customer Pro.

Examples

iex> Mailings.Mailer.send_pro_text("test@example.com")
:ok

send_tp_html(customer)

send_tp_html(String.t()) :: :ok

Send an email in format html for customer TP.

Examples

iex> Mailings.Mailer.send_tp_html"test@example.com")
:ok

send_tp_text(customer)

send_tp_text(String.t()) :: :ok

Send an email in format txt for customer TP.

Examples

iex> Mailings.Mailer.send_tp_text("test@example.com")
:ok