Taxgig v1.0.0-beta.1-0-g3b4f48f+dev Ptin.Services.Downloads

This module contains functions to manipulate files. https://www.irs.gov/pub/irs-utl/FOIA_Extract.zip https://www.irs.gov/tax-professionals/ptin-information-and-the-freedom-of-information-act

Types

error_tuple()

Specs

error_tuple() :: {:error, reason()}

posix()

Specs

posix() :: :file.posix()

reason()

Specs

reason() :: any()

result()

Specs

result() :: success_tuple() | error_tuple()

success_tuple()

Specs

success_tuple() :: {:ok}

Functions

create(attrs)

Specs

create(%{required(atom()) => any()}) :: result()

Insert into table an Expires with data, attrs is expired - "Updated August 30, 2019" and url - source archive or zip "https://www.irs.gov/pub/irs-utl/foia_utah_extract.zip" after couple minutes unpack zip archive and inserted in table an Expires then has been created ptins table and saved data in directory priv/data by path - base_data() from file with extention csv convert to json in bottom line deleted it. Every time when you have created Expires table in DB will be created only one record, previously record will be destroyed after repeat create action, only one record saved in DB.

get(path \\ base_data())

Specs

get(bitstring()) :: result()

Download files, rename and unpack it. Ptin.Services.Downloads.get() or Ptin.Services.Downloads.get("priv/data")

get!(path \\ base_data())

Specs

get!(bitstring()) :: result()

Download file, rename unpack and insert into DB.

get_csv(path)

Specs

get_csv(bitstring()) :: result()

insert(path, file \\ "foia_utah_extract.csv")

Specs

insert(bitstring(), bitstring()) :: result()

Take CSV file and import via stream only custom fields with Ecto.Repo to Database, path - directory by date, file - name CSV file. Ptin.Services.Downloads.import("2020-1-16", "foia_west_virginia_extract.csv")

insert!(path, file \\ "foia_utah_extract.csv")

Specs

insert!(bitstring(), bitstring()) :: result()

Take CSV file and import via stream only custom fields with Ecto.Multi to Repo, path - directory by date, file - name CSV file. Ptin.Services.Downloads.import("2020-1-16", "foia_west_virginia_extract.csv")

remove_file(path, file)

Specs

remove_file(bitstring(), bitstring()) :: result()

Delete only files into directory by date. path - name directory by date, file - name file by delete. Ptin.Services.Downloads.remove_file("2020-1-15", "foia_vermont_extract.csv")

remove_repo(path)

Specs

remove_repo(bitstring()) :: result()

Destroy only directory by date with files. Ptin.Services.Downloads.remove_repo("2020-1-15")