Taxgig v1.0.0-beta.1-0-g056cfdc+dev ServerWeb.GraphQL.Data

A single Dataloader struct can have many different sources, which represent different ways to load data, provides an easy way efficiently load data in batches.

Types

option()

option() :: {:timeout, pos_integer()} | {:get_policy, atom()}

source_name()

source_name() :: any()

t()

t() :: %Dataloader{
  options: [option()],
  sources: %{required(source_name()) => Dataloader.Source.t()}
}

Functions

data()

data() :: t()

query(queryable, params)

query(map(), map()) :: map()