Taxgig v1.0.0-beta.1-0-g056cfdc+dev ServerWeb.HTTPoison.GoogleBehaviour behaviour

Because we are mocking the api requests in ServerWeb.Provider.OauthGoogle we have to have a separate module to delegate the functions we use to the actual HTTPoison module, so that's all we do here.

Callbacks

generate_refresh_token_url()

generate_refresh_token_url() :: String.t()

generate_url()

generate_url() :: String.t()

refresh_token(token)

refresh_token(token()) :: {:ok, map()} | {:error, binary() | map()}

token(code)

token(code()) :: {:ok, map()} | {:error, binary() | map()}

user_email(token)

user_email(token()) :: {:ok, map()} | {:error, binary() | map()}

user_profile(token)

user_profile(token()) :: {:ok, map()} | {:error, binary() | map()}

verify_token(token)

verify_token(token()) :: {:ok, map()} | {:error, binary() | map()}