Taxgig v1.0.0-beta.1-0-gede5592+dev ServerWeb.HTTPoison.S3Behaviour behaviour

AWS S3 mocking the api requests in Core.Uploaders.S3 we have to have a separate module to delegate the functions we use to the actual HTTPoison module, so that's all we do here.

Types

file_spec()

file_spec() :: {:file | :url, String.t()}

Callbacks

put_file(upload)

put_file(upload()) :: {:ok, map()} | {:error, binary() | map()}

put_file(upload()) ::
  :ok | {:ok, file_spec()} | {:error, String.t()} | :wait_callback